requirements.txt
aiogram==3.4.1
loguru==0.7.2
python-dotenv==1.0.1